Module usbd_hid::descriptor

Implements generation of HID report descriptors as well as common reports

Modules

generator_prelude

Prelude for modules which use the gen_hid_descriptor macro.

Structs

KeyboardReport

KeyboardReport describes a report and its companion descriptor that can be used to send keyboard button presses to a host and receive the status of the keyboard LEDs.

MediaKeyboardReport

MediaKeyboardReport describes a report and descriptor that can be used to send consumer control commands to the host.

MouseReport

MouseReport describes a report and its companion descriptor than can be used to send mouse movements and button presses to a host.

SystemControlReport

SystemControlReport describes a report and descriptor that can be used to send system control commands to the host.

Enums

KeyboardUsage

KeyboardUsage describes the key codes to be used in implementing a USB keyboard.

MediaKey

Media player usage ids that can be used in MediaKeyboardReport

SystemControlKey

System control usage ids to use with SystemControlReport

Traits

AsInputReport

Report types which serialize into input reports, ready for transmission.

SerializedDescriptor

Report types where serialized HID report descriptors are available.

Attribute Macros

gen_hid_descriptor

Attribute to generate a HID descriptor & serialization code